requirements.txt
google_api_python_client==2.160.0
google_auth_oauthlib==1.2.1
moviepy==1.0.3
protobuf==4.25.0
python-dotenv==1.0.1
Requests==2.32.3
schedule==1.2.2
google-genai==1.0.0
google-api-core==2.11.0
